Advanced board features implemented.
Two new bitplanes can be used as net input planes: illegal moves for the
current player (with simple ko but not superko) and intersections which
are the last liberty for a group of either player.

These bitplanes are added to the 2 stones bitplanes for the inputmoves,
which are lowered from 8 to 4 to keep the number of planes.

The option --adv_features enables the writing of this new kind of
training chunks.

The reading of the same kind of weights files is always enabled: a
weights file coded with version v=17 (first line) is interpreted as having advanced features coded in the input planes.