Bugfix: numeric robustness problem of integral formula
The integral average of the sigmoid is now a function of namespace Utils and it
is more robust than before. It is called twice: by the method
SearcResult::get_eval_with_bonus() and inside UCTNode::create_children() to
compute m_agent_eval. It has two integration extrema as arguments, so it is
ready for the generalization of the agent with mu.