Integer komi and jigo management. leelaz done, autogtp started.
Now leelaz deals with jigo (draw) at all levels: it can load and
save a RE[0] sgf, it can score appropriately a jigo after 2 passes,
it can dump training file with '0' winner.

UCTSearch is updated accordingly.

The winner is coded as FastBoard::BLACK and WHITE when one player
wins and as EMPTY in the case of jigo. INVAL is used for all other
cases. (Previously EMPTY was used to code games ended by time).

Started implementing jigo on autogtp, for the part dealing with
leelaz. The string m_winner is "jigo" in the case of jigo.

The part of autogtp that communicates with server is to do.