Simmetries exploited!
New line-command option --symm modifies the agent behaviour in the following way.

For every statem the policy probabilities of moves which are equivalent by symmetry are
collapsed and summed onto a single one of them, which is randomly chosen. The other
moves are not even included in the UC tree.

The total number of visits is then split again evenly among equivalen moves at the
moment of writing the training information.