Tensorflow code updated to write "advanced features" weights.
Nothing has to be done about training when passing from
regular LZ nets to "advanced features", since the new input
planes are coded as the other ones and the training code
doesn't need to know shat it represents.

Nevertheless leelaz needs to recognize these weights file as
different, so the version number at the beginning of the file
must be changed.

This can be chosen in config.py