Japanese weak mode completed
Command line option -j or --japanese changes the behaviour of SAI in
several ways.

(Notice that this is NOT intended to be used in training games.)

1) <pass> exploration is enhanced, by perturbing policy at root
with +20% probability. This doesn't change exploration too much
during the early phase of the game, because after a few visits the
UCT does not select pass with high frequency anymore.

2) <pass> move is forbidden unless the agent is confident that the
game is finished.

3) At every move, after exploration, the agent checks if the game
is finished by comparing alpha values of <pass> and <bestmove>.
Comparison is done by computing th median value over explored
subtrees (of size at least EXPLORE_MOVE_VISITS, which defaults to
30 nodes).

This check and what follows heavily relies on the network being
able to understand the board situation. If the net is not good
enough, then inconsistend behaviour may follow.

If <bestmove> is not significantly better than <pass>, then the
presence of dame is checked and in case a dame is present a move is
found that will achieve the shortest sequence of moves yielding any
dame being played. Only "not-losing-point moves" are allowed in
this phase.

4) After all dames have been played, the agent is allowed to pass.

5) The GTP command final_score gives the japanese territorial
score.