Japanese mode tuning
After thourough testing, some changes in the way the search
is done seemed necessary.

Now Japanese mode works fairly well, though it is slow and
could benefit from more optimizations.

One minor issue we are aware is that in the presence of seki
with eyes and opponent's stones inside the eyes, by Japanese
rules those stones should not be countet ad prisoners, while
final_score command counts them. In fact, in that case the
player would capture the stones before passing, and hence the
score is correct. A difference would emerge in the case there
were 2 stones inside the eye, since in that case the player
would capure, getting a large ko threat, the opponent would
place another stone inside the eye and then the player would
capture again. The score would be wrong by 1 in that case.