Experimental change in UCT to decrease pass
Nets around generation 100 of 9x9run01 show a tendency to pass
when losing. The reason seems that even a small number of visits
to second pass is enough to raise winrate from say 1% to 4%.

Modified uct_select_child as to do the minimum number of visits
to losing double pass.