New option --nrsymm
Same as --symm, but the move is chosen to be in the general direction
of the 'polite' eightth of the board, instead of randomly. Disable
NNCache.

In this way if we play many games starting from the initial position,
SAI builds the tree of all the games, collapsing symmetric moves and
with natural variety (disabled NNCache).