Tools for ML-Elo computation from match results
ParseResults is a C++ tool that reads saiXX-matches.xls and
saiXX-netdata.xls files (as dumped by sairesults.sh and sainetsdata.sh)
and produces a file saiXX-nets.csv with nets hashes, games and training
steps steps and a file saiXX-vs.csv with a square matrix X of integers,
were the element x_ij is the number of wins of net i vs j.

Only the nets that are connected to the reference point (random net)
through a path of matches are included.

saivs.py is a Python/Tensorflow tool that computes the most likely Elo
ratings given the match results, by maximum likelihood estimation. It
produces a file saiXX-ratings which is in the same order as the nets in
saiXX-nets.csv