New option --policy_temp
Added new option --policy_temp

It is the same as --softmax_temp, but applied after removal of
illegal moves and after collapsing of symmetrical moves, which
is the correct behaviour. The other option is kept for
compatibility, but the new old should be used instead.

Experimental UCT mod for second pass deactivated in japanese
mode.

NNCache is not even populated when the option --nrsymm is used.