Father id and sgf dump added
Instead of father_bonus and father_base, just father_progid is
exported, as this also allows to deduce node depth.

A comprehensive sgf with the explored tree is also generated.

The filename is automatically generated with main info.

The order of the subtrees is from larger to smaller, instead of
the opposite, since it would be much more difficult to get the
sgf with the main variation in the other case.