Pass tinkering (for handicap games)
Pass has sometimes a very large policy when the net is losing.
Maybe this has to do with the fact that passing in Tromp Taylor
scoring with dead stones has a nonzero probability of winning in
losing positions.

To deal with this problem, which becomes relevanto for example in
handicap games, if -d is not specified, then a first pass is
treated as a second pass in UCT_select: namely final_score is
checked and if losing we set psa=winrate=0.

Moreover fpu reduction is done with respect to the largest winrate
among the parent and the visited children and it is maxed with 0.