Komi dependent policy
New type of network structure introduced.

Code 49 for 110001
            +------- komi_policy
             +------ advanced_features

The komi (from the point of view of current player) is fed into the
policy head, after the 1x1 convolution and before the dense layer.

This is simply done by appending one additional float to the
BOARD_SQUARES*POLICY_OUTPUTS tensor.

leelaz recognizes the new file version automatically and is backward
compatible.

To train a network of this format it is enough to change a variable
in config.py -- in particular the training data are the same as for
the usual advanced_features (code 17) structure.