Bugfix: PASS chosen too often
A recent commit introduced a modified behaviour for UCTNode::uct_select() when
evaluating a PASS move: if the option -d is not given the scores of PASS (psa
and winrate) are modified according to Tromp-Taylor scoring (in a way which is
similar to when this is a *second* pass).

This was wrong in some cases, resulting in a high probability of choosing PASS
(once) when the player was ahead by TT score.