New net structure for komi policy nets
As observed on 2019/01/08 in the chat, with the initial implementation
of komi policy nets, every possible move had just one coefficient that
linked its policy to the komi value. In particular when the komi (say)
increases, there will be a fixed set of moves whose policy would
increase, and a fixed set of moves whose policy would decrease. This
independently of the board position.

The new net structure introduces two layers of 14 units with

(1x1_policy_out,komi) -> L1 -> L2 -> hidden

and then plugs 'hidden' as additional input of the dense layer in the
policy head.

With this commit only training is updated.