BayesElo initial update for draws
The presence of draws in the match results broke the BayesElo
scripts. This first commit corrects ParseResults.cpp so that
it correctly read draws from Mongodb match results dump and
writes one additional file '-num.csv' with the table of number
of games.

More precisely, the first line is the dimension of the square
table and element table[i][j] is the integer number of games
between player i and player j. Clearly the table is symmetric.

The number of draws can be recovered by subtracting

table_num[i][j] - table[i][j] - table[j][i]

Other minor improvements: now pseres takes <saiXX> and the
hook hash from the command line.

The number of header rows in the results file may vary: it was
3 before and is now 4. The program reads lines until it finds
one starting with '{' and complains if the number of rows is
unexpected.