Max 75% of all visits per node
To increase exploration of openings, a hard bound on the number of
visits is introduced, which is equal to 75% since 25% is the weight
of Dirichlet noise.

At the same time, playouts are not stopped when no other move is a
contender.

Both changes apply only when the move number is less than random_cnt.