Median alpkt in game comment
Besides blackevals and node average (agent) winrate evaluation, a
second tree search statistics is added to each node, namely, an
approximation of the sample median of the alpkt values of the subtree
rooted at the node.

The median is approximated by an on-line algorithm.

The value is computed just after that move has been chosen and in
particular it is not updated when a new search starts on that node
for the following move.

The new value is given as the first one in the comments of the move,
when the sgf is printed.

General reorganization of how the search evaluations are stored
inside a GameState, with a specific struct StateEval.

Previously the statistics were computed on the current node after the
search, stored in the GameState of the following move and then the
comment would be printed a move before. In this way the statistics
would include all the visited nodes below current. Now the statistics
are limited to the nodes visited before the move is chosen.

Another difference is that now we do not have any evaluation for the
starting position, instead we print the list of variable names there.

Of course if the only aim was to get the median of alpkt values at
move time, in order to put it into the sgf comment, a proper sample
median could have been computed after the tree search. We prefer here
to use the online version because:

1) In the future it may well be that an on-line median alpkt value
could be used at search time. For example to choose when to pass.

2) Some preliminary experiments with sample median showed uncertain
behaviour and in particular it could be sometimes very different from
the online approximation, and in many of those cases preatty wrong.
Our present understanding is that a) maybe the median computing
recursion has some bugs and b) it seems that computing the median
without using the values of beta can be suboptimal. The online
algorithm implicitly puts more weight on the values that are given
later, which are deeper in the tree and hence have larger values of
beta.