Restrict TT score effect on pass in MCTS
This commit tries to address the problems with passing due to
Tromp-Taylor score evaluation. Its effects are activated with
the new option:

 --restrict_tt

The previous pass-explore logic is removed, whether or not the
above option is used.

Without ad hoc code we can never expect the network to
understand when it is time to pass. Hence we expect pass to
always have very low policy for strong networks, and the
program should in principle resort to passing only when other
moves lose (the game or more points).

(The previously used pass-explore logic would forbid
lose-passing exploration completely, hence making the program
play badly in the endgame.)

Nevertheless Tromp-Taylor scoring may sometimes confuse MCTS,
because a double pass evaluation will usually be very different
from the real one, because of dead groups not being removed.

A player who is losing may see a small probability of winning
when passing, if the opponent also explores pass and TT score
is favorable. This player will often start passing in the hope
that the opponent also passes, hence losing more points.

To correct this behaviour, the MCTS is changed in the following
ways:

1) The previous pass-explore logic is removed.

2) TT score is always stored in the tree node of the second
   pass, where it belongs. In this way the program always
   knows the true (Tromp-Taylor) consequences of a second pass
   and can choose accordingly whether it likes the move or not.

3) TT score is never used to update ancestor nodes that precede
   the first pass, since it is not the true (Chinese) score of
   the game and should not have effect on choosing between
   non-pass moves. Those nodes are instead updated with the raw
   network evaluation of the Chinese score at the second pass
   node.

4) When the first pass node evaluation has to be updated with
   the value of its PASS child (that is, the second pass),
   initially we use raw network evaluation as for the preceding
   nodes, and after the number of visits is reasonably high we
   start using TT score. In this way, if the opponent would
   actually like the second pass, we know this and can use this
   information to choose whether we like the first pass.