Option to disable NNcache
Added new option --nocache to disable NNcache. This is now
independent from the option --nrsymm, which previously would
also disable cache and does not anymore.

Also now if --nocache is used, the cache is still created, but
kept as small as possible. (10 entries or 1% memory.)

Some small errors corrected in the calling of get_output().