Chain liberties feature planes
Added another optional type of input planes. For each
intersection of the board, these planes encode the number of
liberties of the chain/string (largest set of connected stones of
the same color) containing the stone in that intersection.

This is a non-local information that is expected to improve the
network understanding of board positions.

The encoding by default uses 4 bit planes: 1, <=2, <=3, <=4
liberties.

The order of the planes is: own stones, opponent stones,
[advanced features], [chain liberties].

These planes are used while reading the network if the version is
'81' (for now; the plan is to encode in version number bits 0...7
the following flags: 0-1 for LZ/ELF, 2-3 unused, 4 adv_features,
5 komi_policy, 6 chain_libs, 7 chain_size).

These planes are written in training data with the option

  --chainlibs_feat