Fix for 'dumbpass'
Occasionally with dumbpass a network would make a wrong first pass,
losing by not removing dead stones, because it did not consider the
opponent answer to pass.

This behaviour is unwanted, in particular with --restrict_tt option.

To correct this, when the first pass is played in the MCTS the second
pass has the policy raised by 0.2, so that the second pass is considered
as one of the first moves.