Rich training data
Now leelaz writes additional info inside training data, and
the training script is updated to work with these (while still
being compatible with previous formats).

On the 'winner' line, three more fields are added; these are
floats space-separated, and represent statistics computed on
the explored UCTree rooted at that position:

alpkt_median_online - online approximated median of alpkt values
beta_median         - sample median of beta values
azwinrate_avg       - average of net raw winrates of nodes
                      (current komi, lambda=mu=0, 'agent zero')

On the 'side to move' line, in addition to the komi, two more
fields are added, namely the sgfhash (sha256sum of the sgf
string) and the move number (starting from 0). In this way it
is now possible from training data to track back the game that
generated it. In fact the sgfhash is the same present in the
server database and it can also be used directly to see the
game on a browser with /view/sgfhash?viewer=wgo

To get the right hash, the sgf must be fixed with respect to
the opponent player string and the comment starting GTP
commands, in exactly the same way as autogtp does. This is
presently done with a quick patch that is not very robust;
hence some care should be put on future modifications in the
code of autogtp.