Bayes GOR for pruned networks
ParseResults (pseres) is usually invoked with the flag '-p' which
tells to prune networks that have had only one opponent in the matches
(leaf of the graph) and networks that lost or won all the games in the
matches with more than one opponent.

Then one calls the python/tensorflow script (saivsdraws.py) which
computes the maximum likelihood estimates of the ratings of the
remaining nets.

Then, by calling again ParseResults, a new file saiXX-rated-nets.csv
is written, with estimates of the rating of all connected nets, based
on the other ratings.