Better pass logic and diagnostic tools
A problem was found with --restric_tt option: sometimes when the Chinese
score is near 0 and the Tromp-Taylor score is 0, the "winning" player
would pass instead of continue playing by capturing dead stones.

This unwanted behaviour suggested to improve the available diagnostic
tools, to better understand MCTS tree exploration in the critical
examples.

Subsequent analysis showed that the option was not quite affecting the
exploration of the pass but rather of its alternatives, because the
imprecision of the net evaluation propagated in nodes above passes could
be large.

The simple correction adopted here is to propagate in nodes above the
first pass the same result that would be stored in the first pass move,
i.e. net evaluation for the first playouts and then Tromp-Taylor score
when the second pass has a large share of its father visits.

As for the diagnostics:

'eval' command was improved by including all the visits to second pass
and not just the first one and by adding the fields

  first_move:  the generation 1 ancestor of this node, i.e. the first
               move of this playout;

  eval_update: the value used for updating the first_move node, which in
               the case --restrict_tt may be different from the field
               net_eval.

DEBUG compile mode was improved in the tree exploration output, which
now shows playouts with moves, visits, value for updating and eval after
the update.