New random moves sampling, sgf comment.
If the user specifies -m <num>, then up to num random moves are chosen
during play, but not necessarily the first ones.

For each move a card is extracted (with replacement) from a num-sized
deck. If the card was never extracted before, the move is chosen
randomly, otherwise the best move is chosen.

This way also moves in the middle and end of the game can occasionally
be random.

The usual controls apply, so if blunders are not allowed, or if the
allowed blunders are already spent, then the random move is chosen among
those with winrate near enough to the best move.

The sgf comments are modified in that the last field in no more a single
bit for blunder/non-blunder, but a bitfield: bit 0 random move, bit 1
blunder move. So for now we have:

0 - nonrandom
1 - random nonblunder
3 - blunder

(Blunders are always random.)

Added display of the move number in normal mode