Beta value amplification coefficient (training only)
This is a bugfix with backward compatibility concerns.

New command line argument --betascale <x> for parse.py which
defaults to 2. The value of x is a float.

The trained nets will correctly estimate beta with x=1.
Otherwise they will estimate x*beta. Since current versions of
SAI do not allow to apply a correction factor inside the main
program, we allow the networks to be trained with an inherent
factor.

This factor was 2 from the beginning because of a bug.

As of first 19x19 run and generation 260, we are still training
with factor 2, because a sudden change will modify the
exploration and hence the training data produced by new nets
self-plays would be discontinuous.