Default options upgrade
Before we were trying to keep the default working mode of SAI
identical or almost identical to Leela Zero.

Since many users don't really check the options and SAI is now
well known in the community, it is preferable to have good
playing options as default.

--nrsymm is now the default (option kept for compatibility)
--nosymm added to get LZ behaviour

--recordvisits now default (option kept for compatibility)
--norecordvisits added to get LZ behaviour

Changed default komi from 7.5 to 7