Bugfix: alpkt_online_median was capped
This commit fixes a small bug: when the online estimator of
alpkt median is updated, the number of visits may be already
increased to 2 and in that case the starting value is 0 instead
of the first element of the sample.

Starting from 0 and considering the cap for the difference at
0.5, the final value of the estimator becomes capped at about 23
+ 0.5 * ln(visits) in absolute value, which is wrong for large
absolute values of the komi.

Notice that alpkt_median is not used by SAI for playing, but it
is the first number written inside comments, and it is used for
generating branches.