Pass agreement improvement
Range is now dependent also on the winrate. As before, if the winrate is 90:10
or 10:90, a confidence of 90% on the score (through the sigmoid) is required to
pass, but if the winrate is more or less decided this changes. For example a
winrate of 50:50 requires the highest confidence on the score of 99%.

Corrected number of digits in printed information and minimum LCB.

The implemented formula is

         conf = 1 - alpha/2
  -log(alpha) = b * (a/b)^(y/c)

y =  log(w/(1-w)         log-odds winrate
a = -log 0.1             normal confidence
b = -log 0.01            highest confidence
c =  log (0.9 / 0.1)     log-odds normal confidence