pass_agree improvement
Reintroduced a (bland) verification that the player is agreeing to pass and
lose if he believes he's winning.

Default is still pass_agree=false, but also --restrict_tt set it to true,
so recent client will use this option on self-plays even before all clients
are required to be compatible with the option.