Resigning aware of agent
With lambda or mu positive, the agent evaluation of the winrate is
regressed towards 0.5, so resigning happens much later than what is
expected.

Now we transform resign threshold according to the values of lambda
and mu to get a more reasonable value. The transformed threshold is
printed when resigning happens.