Autogtp bugfix: buffer White comments to file
Previously White's sgf string with comments was taken from the stdout of
sai process via a simple "printsgf" command (without arguments).

This from time to time resulted in a string truncated to 8194 characters.
No reason was found for this bug, and the behaviour itself was erratic.

Now the sgf is instead buffered to a file like Black's one, using a
"printsgf <filename>" command.