Improvement in online alpkt estimation
It seems better to estimate the "fair komi" value after MCTS with
this new method. Instead of approximating the median of alpkt in
subtrees, which had no robust justification, we consider 'w', the
average of the whole winrate function of the nodes in the subtree.

We try to estimate a value y such that w(y)=0.5.

The chosen algorithm can be shown to converge to the correct value
under suitable hypothesis, for the tree size going to infinity.
More importantly, several experiments performed with real positions
showed that this method works very well, while the previous one had
poor performances.

The theoretical coefficient should have been 4, as the derivative
of the sigmoid in 0 is 1/4, but actually 5 seems to work a little bit
better in the experiments.