Bayes Elo upgrade: Gaussian prior
Previously Elo estimation was exclusively done by Maximum Likilihood.

Now, when two nets are near (in particular every time a net is
matched against a single other net) a Gaussian prior is assumed for
the Elo rating difference (with mean 0 and standard deviation 25 Elo
points) and estimation is done by maximizing the joint probability of
Elo difference and match outcome.

We also dropped the bound for too large difference, as it is not
needed anymore (and was probably wrongly implemented in the first
place).