Resign with agent fine tuning
There was an unwanted behavior causing SAI to resign incorrectly, in
particular when the agent is used with large values of lambda and/or
mu (for example with lambda=1 and mu=0.5 this happens every time).

This is caused by the fact that the agent winrate which is averaged
on the tree is the one computed on the interval of komi values given
by the agent parameters. Since this interval depends on the raw
network evaluation at the root node, it is often imprecise.

To decide whether it is time to resign usually we compare the agent
winrate with a transformed threshold and this would cause erratic
mistakes with large values of lambda and/or mu.

To correct, before resigning, we also check that the raw newtwork
winrate at current komi is below the resign threshold.

Added disply of the agent integration interval (relative to the real
komi and from black point of view) before the tree analysis.