Elo script better output
Previously the script would output the total loglikelihood.

Now it prints average loss and excess average loss with respect to the
theoretical Elo model.

Remark: theoretical model is not exact, because it requires the winrate to
depend only on the difference in Elo ratings. This would for example exclude
rock-paper-scissors situations, which in fact happen quite often.

A better model would use vector ratings, with dimension larger than 1.