Detailed agent
A. A more detailed version of SAI agent is implemented.  This
agent has 8 parameters: four different couples (lambda, mu) for
winrate above/below 0.5 and for each of the players.

To this end SAI has now a concept of what color it should be
playing, in order to be able to consider different agent
parameters for itself and for its virual opponent in the MCTS.

The typical use is to give positive parameters for SAI, and zero
for the virtual opponent, so that in the MCTS SAI does simulate
an opponent that plays just to win.

Moreover there is the flexibility to give different set of
parameters to use when winrate is above or below 0.5.

The first 'genmove', 'go', 'play' or 'auto' command fix what
is SAI/CPU color. (In the same order: same, same, other, both.)

The command line arguments --lambda and --mu now accept one to
four numbers each, comma separated (no spaces). The order is:
(1) CPU ahead, (2) CPU behind, (3) human ahead, (4) human behind.

If less than four numbers are given,
(4) := (2); (3) := (1); (2):= (1).

The same list of numbers may be given to 'lz-setoption' command
with names 'lambda' and 'mu'.

Agent's parameters lambda and mu are determined for each node and
updated at each simulation by UCTNode::set_lambda_mu(), according
to the node's statistics and the corresponding GameState. For
this implementation, UCTNode::get_raw_eval() and
GameState::is_cpu_color() are used, but more general methods may
be implemented.

B. New 'lz-search_reset' command

The 'lz-setoption' 'lambda' and 'mu' subcommands will *not* reset
the search tree anymore, but a specific command 'lz-search_reset'
is available to do that.

C. Bugfixes

The method UCTSearch::reset() was wrong. Removes and replaced.
In UCTSearch::dump_stats() there was a myprintf() with the wrong
number of arguments.