Bayes Elo computation refactoring
The new procedures can be used keeping fixed the ratings of networks
older than a given hook.

The method is based on creating a connection graph to the hook based on
relevant matches in order to drop the disconnected networks and
to discriminate between leaf and trunk ones.

Matches between two nets that are both older than the hook or have zero games
are ignored while computing the connection graph.

The trunk networks' ratings are retrieved from the existing ratings file
from previous runs.

The ratings of the trunk networks younger than the hook is established via
a max likelyhood fitter, while the hook and its older networks' have
their rating kept fixed.