Better resigning in handicap games
Previously the resign threshold was tweaked considering handicap
stones only when it was left to its default value.

Now it is also when a different threshold is given. Hence, the user
does not need to adapt resign threshold to handicap games.

Moreover, SAI with a SAI network estimates the relative strength of
its opponent, and if it sees that there is the possibility to recover
points before the 80% of 361 moves are played, it will play on.