FPU-average
We replace fpu reduction with a different approach: fpu for unvisited
nodes is the average of the winrate among all siblings.

The average is not weighted with the number of visits. The best child
is excluded, apart for the case in which no other child has visits.

The sense is the following: after the first child is visited, fpu is
equal its winrate, so a second child is visited immediately.

If the winrate of the visited children is similar, it is expected
that other children will have similar winrate, and so fpu is high
and many moves are visited at least once.

When several lower winrate children start to decrease the average
winrate, fpu becomes lower and new children are not visited as
often.

In the case that there is only one reasonable move, there will be
exactly one visit to a second child.

This new fpu approach has been tested estensively, with different
parameters and different networks, and it seems to give an average
improve of about 5% wins. In no condition it showed to be worse
than the previous. As such, it is now the default.

The command line argument --fpu_old reverts to previous behaviour
(which was already a little different from LZ's we remark).