Detailed agent fix
Previously an initial 'play' command would fix the role of CPU as the
other player. This was not right, as Go programs typically use a
sequence of 'play' commands to get to a specific position.

Now only 'genmove', 'go' and 'auto' commands will fix CPU role.

Moreover, if the agent parameters for CPU and human are all equal, then
the role is fixed anyway, but the warning messages are suppressed.