Fix score and komi with handicap
A previous commit broke the computation of area score in the case
of handicap games, because of a wrong understanding: handicap stones
should not count for area score, or alternatively the amount of stones
needs to be added to the komi.

Because of this, also SAI behaviour needed to be changed in several places,
otherwise the score assumed by SAI would have been wrong.

Finally, this affects also LZ networks, that are restricted to komi 7.5, as
they would also assume the wrong score in handicap games (if the user was not
aware of this).

To be more explicit, if a LZ network is used in a handicap game, the program
will always play as if the komi was 7.5 for White and the handicap stones were
points just like the other black stones. If for example there was a 5 stones
handicap, LZ would play as if the komi was 7.5 - 5 = 2.5.

To account for this behaviour, when the network is LZ and not SAI, every time
the handicap is changed, the komi is reset to 7.5 - handicap_stones, because
that is the true komi that LZ is assuming in that case.